% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGenerator.R
\name{eGenerator}
\alias{eGenerator}
\title{Generate enhancer locations}
\usage{
eGenerator(geno, maxgap = 10^4)
}
\arguments{
\item{geno}{an n*p genotype matrix, where n is the sample size and p is the number
of rare varints included.}

\item{maxgap}{threshold to split independent loci (default=10^4)}
}
\value{
The function returns a data frame containing generated enhancer location.
The first column is serial number of the enhancer.
The next two columns are physical location of the enhancer.
The nest two columns are adjusted location of the enhancer 
(the position in the \code{geno} matrix).
The last column is enhancer length.
}
\description{
The \code{eGenerator} function generates regulatory rgeion locations 
if not specified by users.
}
