% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSCAN_search.R
\name{eSCAN_SEARCH}
\alias{eSCAN_SEARCH}
\title{eSCAN_SEARCH}
\usage{
eSCAN_SEARCH(genotype, nullmod, new_enloc, fam = 0, weights)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix)}

\item{nullmod}{an object from fitNullModel() function}

\item{new_enloc}{an m*3 matrix of enhancer information}

\item{fam}{a logical value to indicate phenotype family: 1 indicates binomial, 0 indicates gaussian}

\item{weights}{vector of weights for variants of interest}
}
\value{
a vector with len=n_en, listing p-value for each enhancer
}
\description{
compute p-value for each enhancer
}
