% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSCAN_thres_cpp.R
\name{eSCAN_THRES}
\alias{eSCAN_THRES}
\title{eSCAN_THRES}
\usage{
eSCAN_THRES(genotype, nullmod, new_enloc, fam = 0, times, weights)
}
\arguments{
\item{genotype}{n*p genotype matrix}

\item{nullmod}{fitted null model from GENESIS package}

\item{new_enloc}{NULL or matrix whose number of rows is enhancer number, and has 6 columns, chr No., start_pos, end_pos, start_adj, end_adj, length}

\item{fam}{logical value, 1 indicates binomial phenotype, 0 indicates gaussian}

\item{times}{integer to indicate B in Monte Carlo simulation}

\item{weights}{vector of weights for variants of interest}
}
\description{
compute empirical threshold
}
