% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_enhancers.R
\name{generate_enhancer}
\alias{generate_enhancer}
\title{generate enhancer}
\usage{
generate_enhancer(geno, maxgap = 10^4)
}
\arguments{
\item{geno}{genotype matrix}

\item{maxgap}{threshold to split independent loci, with default value 10^4}
}
\value{
a data frame involving ennhancer information
}
\description{
generate enhancer list for further use
}
