% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing}
\usage{
preprocess(geno, enhancer, gap)
}
\arguments{
\item{geno}{an n*p genotype matrix, where n is the sample size and p is the number
of rare varints included.}

\item{enhancer}{a data frame of annotation information with dimension q*6, where
q is the number of the target regulatory regions. The six columns indicates no., physical
start position, physical end position, start position in the \code{geno} matrix, 
end position in the \code{geno} matrix and region length, respectively. 
If annotation information is not specified by users (default=NULL), a data frame of 
enhancer location will be automatically created by \code{\link{eGenerator}} function.}

\item{gap}{threshold to split independent loci (default=10^4)}
}
\value{
The functoin returns a list with the following elements:
\code{genotype}: An n*p genotype matrix by increasing order of variant position.
\code{MAF}: A vector (length=p) of the minor allele frequency.
\code{new_enloc}: A data frame of enhancer location of dimension q*6, where
q is the number of the target regulatory regions. The six columns indicates No., physical
start position, physical end position, start position in the \code{genotype} matrix, 
end position in the \code{genotype} matrix and region length, respectively.
}
\description{
This function is a preliminary step for subsequent analysis using eSCAN.
}
