% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSCAN.R
\name{eSCAN}
\alias{eSCAN}
\title{eSCAN}
\usage{
eSCAN(
  genotype,
  nullmod,
  new_enloc = NULL,
  gap = 10^4,
  times = 1000,
  alpha = 0.05,
  analy = FALSE
)
}
\arguments{
\item{genotype}{n*p genotype matrix}

\item{nullmod}{fitted null model from GENESIS package}

\item{new_enloc}{NULL or matrix whose number of rows is enhancer number, and has 6 columns, chr No., start_pos, end_pos, start_adj, end_adj, length}

\item{gap}{threshold to split independent loci, with default value 10^4}

\item{times}{simulation times for MC, with default value=1000}

\item{alpha}{significance level}

\item{analy}{TRUE indicates analytical threshold, FALSE indicates empirical threshold}
}
\value{
a list containing 'res', 'res0', 'thres_mat', 'thres'
}
\description{
scan the enhancer regions
}
