% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnull.R
\name{fitNull}
\alias{fitNull}
\title{fitNull}
\usage{
fitNull(x, outcome = NULL, covars = NULL, fam = "gaussian")
}
\arguments{
\item{x}{a data frame containing outcome variable and covariates}

\item{outcome}{a character string specifying the name of outcome variable in x}

\item{covars}{a vector of character strings specifying the name of covariates in x}

\item{fam}{a description of the error distribution and link function to be used in the model. Can only be either "gaussian" for continuous phenotype or "binomial" for binary phenotype}
}
\value{
The function returns an object of model fit from \code{fitNullModel}. See \code{fitNullModel} in GENESIS package for more details.
}
\description{
Fit generalized linear model under the null hypothesis. It's a wrapper of fitNullModel from GENESIS package.
}
