% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing}
\usage{
preprocess(geno, enhancer, gap)
}
\arguments{
\item{geno}{an n*p genotype matrix, where n is the sample size and p is the number
of rare variants included.}

\item{enhancer}{a data frame of annotation information with dimension q*6, where
q is the number of candidate regulatory regions. The six columns indicate index, 
start position, end position, start index, end index
the \code{geno} matrix sorted by genomic positions and length of the enhancers, respectively. 
If annotation information is not specified by users (default=NULL), a data frame of 
locations of the enhancers will be automatically created by \code{\link{eGenerator}}.}

\item{gap}{threshold to split independent loci (default=10^4).}
}
\value{
The function returns a list with the following elements:

\code{genotype}: an n*p genotype matrix sorted by genomic positions.

\code{MAF}: a vector (length=p) of minor allele frequencies.

\code{new_enloc}: a data frame of locations of the enhancers of dimension q*6, where
q is the number of candidate regulatory regions. The six columns indicate index, 
start position, end position, start index, end index 
in the \code{genotype} matrix sorted by genomic postions and length of the enhancers, respectively.
}
\description{
This function is data preparation for subsequent analysis using eSCAN.
}
