% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(geno, enhancer, gap)
}
\arguments{
\item{geno}{n*p genotype matrix}

\item{enhancer}{NULL or matrix whose number of rows is enhancer number, and has 6 columns, chr No., start_pos, end_pos, start_adj, end_adj, length}

\item{gap}{threshold to split independent loci, with default value 10^4}
}
\value{
a list containing 'genotype', 'MAF' and 'new_enloc'
}
\description{
pre-process data for eSCAN
}
